+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * `attr::none()` for conditional rendering of an attribute
 * implement `Default` for `Element` and `Attribute`
-
+* `attr::disabled()` and `attr::autofocus()`
 
 ## [1.4.0] - 2024-12-13